requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.119.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
fonttools==4.60.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
narwhals==2.8.0
numpy==2.2.6
openai==2.3.0
openpyxl==3.1.2
packaging==25.0
pandas==2.3.3
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==12.0.0
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
PyMuPDF==1.26.5
pyparsing==3.2.5
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
psycopg2-binary==2.9.11
pytz==2025.2
requests==2.32.5
scikit-learn==1.7.2
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
Werkzeug==3.1.3
WTForms==3.2.1
